feat: userProfile 상태를 DB 정보로 갱신 및 프로필 로직 개선
