+++ eatory-vueproject/src/components/user/UserProfile.vue
@@ -1,7 +1,7 @@
 <template>
-  <div class="profile-container">
+  <div class="profile-container" v-if="userProfile && userProfile.username">
     <!-- 사용자 이름 -->
-    <h1>{{ userProfile.username }}</h1>
+    <h1>{{ userProfile.name }}</h1>
     <!-- 프로필 이미지 -->
     <img :src="userProfile.profileImage" alt="Profile Image" class="profile-image" />
     <!-- 피드, 팔로워, 팔로잉 카운트 -->
@@ -16,7 +16,7 @@
     import Allergy 
     <allergyView/> -->
     <div class="allergies">
-      <span v-for="allergy in userProfile.allergies" :key="allergy.allergyName" class="allergy-item">
+      <span v-for="allergy in userProfile.allergies" :key="allergy" class="allergy-item">
         {{ allergy }}
       </span>
       <button @click="addAllergy">+</button>
@@ -28,48 +28,36 @@
       <div>Height {{ userProfile.height }}  cm</div>
     </div>
   </div>
+  <div v-else>
+    <p>Loading Profile ...</p>
+  </div>
 </template>
 
 <script setup>
-import { ref, onMounted } from "vue";
-import axios from "axios";
+import { computed, onMounted } from "vue";
+import { useUserStore } from "@/stores/userStore";
 
-//props로 전달된 userId를 가져옴
-const props = defineProps(["id"]);
+// userStore 사용
+const store = useUserStore();
 
-// 사용자 프로필 데이터
-// const userProfile = ref({});
+// userProfile 상태를 계산 속성으로 가져옴 
+const userProfile = computed(() => store.userProfile);
+console.log("현재 userProfile 상태:", userProfile.value);
 
-// 사용자 프로필 데이터를 저장할 상태
-const userProfile = ref({
-  username: "",
-  profileImage: "",
-  postCount: 0,
-  followerCount: 0,
-  followeeCount: 0,
-  allergies: [],
-  height: 0,
-  weight: 0,
-});
-
-// API 호출
-onMounted(async () => {
-  try {
-    
-    const response = await axios.get(`/api-user/${props.id}`);
-    console.log("받은 userId:", props.id); //디버깅용 로그 //URL에서 전달된 :id 값 확인 
-    console.log(response.data);
-    userProfile.value = response.data;
-  } catch (error) {
-    console.error("Error fetching user profile:", error.response || error);
-    alert("프로필 정보를 가져오는 중 문제가 발생했습니다.");
+onMounted(() => {
+  if(store.loginUser) {
+    store.getUserProfile(store.loginUser.id).catch((error) => 
+    console.error("프로필 로딩에 에러 발생", error)
+  );
   }
-});
+})
 
-// 알러지 추가 (구현 예정)
+//알러지 추가 기능 (구현 예정)
 const addAllergy = () => {
   alert("알러지 추가 기능은 아직 구현되지 않았습니다.");
 };
+
+
 </script>
 
 <style scoped>
